pyproject.toml
[tool.poetry]
name = "curso-flaskadmin"
version = "0.1.0"
description = ""
authors = ["Tiago Tardelli <tiagob.tardelli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
Flask-Admin = "^1.6.0"
Flask-Login = "^0.6.2"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
